Rename package model -> ui
