+++ src/test/java/ui/pages/OrderPage.java
@@ -1,4 +1,4 @@
-package model.pages;
+package ui.pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
